docling_processor/requirements.txt
google-cloud-storage
google-cloud-firestore
docling
markitdown
functions-framework
pypdfium2
